% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{D31}
\alias{D31}
\title{The Artificial Data of Consisting of as Many as 31 Randomly Placed Gaussian Clusters}
\description{
This artificial data was generated to show the strength of SUPC.
Clustering \code{D31} dataset is difficult for the partition type of clustering algorithms that require an initial set.
However, SUP correctly identifies the 31 major clusters.
}
\references{
Veenman, C. J., M. J. T. Reinders, and E. Backer. 2002. A Maximum Variance Cluster Algorithm. IEEE Trans. Pattern Analysis and Machine Intelligence 24 (9): 1273–80.
}
