% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dist.parallelization}
\alias{dist.parallelization}
\title{Configure how many cores will be used to calculate the distance matrix}
\usage{
dist.parallelization(i)
}
\arguments{
\item{i}{integer.}
}
\value{
\code{NULL}. The function is called for side effects.
}
\description{
Only affect \code{\link[amap]{Dist}}.
}
