% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{plot.supc}
\alias{plot.supc}
\title{Draw related plot of the clustering result}
\usage{
\method{plot}{supc}(x, type = "heatmap", ...)
}
\arguments{
\item{type}{character value. \describe{
 \item{\link[stats]{heatmap}}{Draw a Heat Map of the centers}
}}

\item{...}{other parameters to be passed through.}
}

