% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{supc1}
\alias{supc1}
\title{Self-Updating Process Clustering}
\usage{
supc1(x, tau = 3.5, t = 0.75, tolerance = 1e-04)
}
\arguments{
\item{x}{matrix.}

\item{tolerance}{numeric value.}

\item{parameters}{list.}

\item{implementation}{either \code{"R"} or \code{"cpp"}.}
}
\value{
\code{supc1} returns an object of \link{class} "supc".

\item{One}{First item}
\item{Two}{Second item}
}
\description{
TODO
}
\details{
TODO
}
\examples{
print("hello example")
}

