% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dist.mode}
\alias{dist.mode}
\title{Configure which package is used to compute the distance matrix}
\usage{
dist.mode(mode = c("stats", "amap", "gputools"))
}
\arguments{
\item{mode}{either \code{"stats"}, \code{"amap"}, or \code{"gputools"}.}
}
\description{
Configure which package is used to compute the distance matrix. Note that the fastest implementation depends on the data and machine.
}
