% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{freq.poly}
\alias{freq.poly}
\alias{freq.poly.default}
\alias{freq.poly.dist}
\title{Plot the frequency polygon of pairwise distance}
\usage{
freq.poly(x, ...)
}
\arguments{
\item{x}{either dist object or matrix.}

\item{...}{other parameters to be passed through to \code{\link[graphics]{hist}}.}
}
\description{
Plot the frequency polygon of the pairwise distance.
}
