% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{freq.poly.supc}
\alias{freq.poly.supc}
\alias{freq.poly.subclist}
\title{Plot the frequency polygon of pairwise distance}
\usage{
\method{freq.poly}{supc}(x, ...)
}
\arguments{
\item{x}{either dist object or matrix.}

\item{...}{other parameters to be passed through to \code{\link[graphics]{hist}}.}
}
\description{
Plot the frequency polygon of the pairwise distance. The red dashed line is the used parameter \eqn{r}.
}
