% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{plot.supc}
\alias{plot.supc}
\title{Draw plots of the clustering result}
\usage{
\method{plot}{supc}(x, type = "heatmap", ...)
}
\arguments{
\item{x}{\code{supc} object to plot.}

\item{type}{character value. \itemize{
 \item{\code{"heatmap"}}{draw a heatmap to show the result of clustering}
}}

\item{...}{other parameters to be passed through.}
}
\description{
General function to draw plots for analysis
}
