% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supc1.R
\name{supc1}
\alias{supc1}
\title{Self-Updating Process Clustering}
\usage{
supc1(x, r = NULL, rp = NULL, t = c("static", "dynamic"),
  tolerance = 1e-04, drop = TRUE, implementation = c("cpp", "R"),
  verbose = FALSE)
}
\arguments{
\item{x}{matrix. Each row is an instance of the data.}

\item{r}{numeric vector or \code{NULL}. The parameter \eqn{r} of the self-updating process.}

\item{rp}{numeric vector or \code{NULL}. If \code{r} is \code{NULL}, then \code{rp} will be used. 
The corresponding \code{r} is the \code{rp}-percentiles of the pairwise distances of the data. 
If both \code{r} and \code{rp} are \code{NULL}, then the default value is \code{rp = c(0.0005, 0.001, 0.01, 0.1, 0.3)}.}

\item{t}{either numeric value, function, or one of \code{"static" or "dynamic"}. The parameter \eqn{T(t)} of the self-updating process.}

\item{tolerance}{numeric value. The threshold of convergence.}

\item{drop}{logical value. Whether to delete the list structure if its length is 1.}

\item{implementation}{eithor \code{"R"} or \code{"cpp"}. Choose the tool to calculating result.}

\item{verbose}{logical value. Whether to show some messages during computing or not.}
}
\value{
\code{supc1} returns a list of objects of \link{class} "supc".

Each "supc" object contains the following elements:
\item{x}{The input matrix.}
\item{d0}{The pairwise distance matrix of \code{x}.}
\item{r}{The value of \eqn{r} of the clustering.}
\item{t}{The function \eqn{T(t)} of the clustering.}
\item{cluster}{The cluster id of each instances.}
\item{centers}{The center of each cluster.}
\item{size}{The size of each cluster.}
}
\description{
TODO
}
\details{
TODO
}
\examples{
print("hello example")

}
\references{
Shiu, Shang-Ying, and Ting-Li Chen. 2016. "On the Strengths of the Self-Updating Process Clustering Algorithm." Journal of Statistical Computation and Simulation 86 (5): 1010–1031. doi:10.1080/00949655.2015.1049605. \url{http://dx.doi.org/10.1080/00949655.2015.1049605}.
}
